Rendering Questions in Accordion Component
The proper folder structure was created for the Accordion component:
"src/components/accordion".

In the folder the following files were defined:
--> index.jsx: where the Accordion component is defined
--> styles.css: where the styles for the Accordion component will be added.
--> data.js : where some dummy data resides.

In the component definition, a conditional rendering was performed and in the case
that there are data to be operated on, the question and the collapsable icon are
rendered. Otherwise, "No data was returned" is returned.